cdnでbootstrapを導入 navbarをスタイリング　削除の権限を制限

+++ src/AppBundle/Controller/ChannelController.php
@@ -129,9 +129,11 @@ class ChannelController extends Controller
         throw $this->createNotFoundException('No channel found for name: '.$name);
       }
 
-      $em = $this->getDoctrine()->getManager();
-      $em->remove($channel);
-      $em->flush();
+      if ($this->getUser()){
+        $em = $this->getDoctrine()->getManager();
+        $em->remove($channel);
+        $em->flush();
+      }
 
       return $this->redirectToRoute('channel_index');
     }
